moved DecodableColumnHandle from SPI to UTILS
