+++ presto-redis/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.facebook.presto</groupId>
         <artifactId>presto-root</artifactId>
-        <version>0.106-SNAPSHOT</version>
+        <version>0.111-SNAPSHOT</version>
     </parent>
 
     <artifactId>presto-redis</artifactId>
