+++ presto-redis/src/main/java/com/facebook/presto/redis/RedisColumnHandle.java
@@ -14,7 +14,7 @@
 package com.facebook.presto.redis;
 
 import com.facebook.presto.spi.ColumnMetadata;
-import com.facebook.presto.spi.DecodableColumnHandle;
+import com.facebook.presto.utils.decoder.DecodableColumnHandle;
 import com.facebook.presto.spi.type.Type;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
