+++ presto-utils/src/main/java/com/facebook/presto/utils/decoder/json/ISO8601JsonFieldDecoder.java
@@ -13,7 +13,7 @@
  */
 package com.facebook.presto.utils.decoder.json;
 
-import com.facebook.presto.spi.DecodableColumnHandle;
+import com.facebook.presto.utils.decoder.DecodableColumnHandle;
 import com.facebook.presto.utils.decoder.FieldValueProvider;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.annotations.VisibleForTesting;
