+++ presto-utils/src/main/java/com/facebook/presto/utils/decoder/json/MillisecondsSinceEpochJsonFieldDecoder.java
@@ -14,7 +14,7 @@
 package com.facebook.presto.utils.decoder.json;
 
 import com.facebook.presto.utils.decoder.FieldValueProvider;
-import com.facebook.presto.spi.DecodableColumnHandle;
+import com.facebook.presto.utils.decoder.DecodableColumnHandle;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableSet;
