+++ presto-utils/src/main/java/com/facebook/presto/utils/decoder/raw/RawFieldDecoder.java
@@ -14,7 +14,7 @@
 package com.facebook.presto.utils.decoder.raw;
 
 import com.facebook.presto.utils.decoder.FieldDecoder;
-import com.facebook.presto.spi.DecodableColumnHandle;
+import com.facebook.presto.utils.decoder.DecodableColumnHandle;
 import com.facebook.presto.utils.decoder.FieldValueProvider;
 import com.facebook.presto.spi.PrestoException;
 import com.google.common.base.Splitter;
