+++ presto-utils/src/test/java/com/facebook/presto/utils/decoder/json/TestMillisecondsSinceEpochJsonFieldDecoder.java
@@ -13,7 +13,7 @@
  */
 package com.facebook.presto.utils.decoder.json;
 
-import com.facebook.presto.spi.DecodableColumnHandle;
+import com.facebook.presto.utils.decoder.DecodableColumnHandle;
 import com.facebook.presto.utils.decoder.DecoderTestColumnHandle;
 import com.facebook.presto.utils.decoder.FieldDecoder;
 import com.facebook.presto.utils.decoder.FieldValueProvider;
